Modularize local_gen functions
